hotel_booking_system/api_gateway_service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
SQLAlchemy
SQLAlchemy-Utils
pyjwt
psycopg2-binary
passlib[bcrypt]
python-multipart